+++ package.json
@@ -21,6 +21,7 @@
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-overflow-indicator": "^0.2.2",
     "react-toastify": "^9.1.2",
     "sharp": "^0.32.1",
     "styled-components": "^5.3.10",
